backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-mistralai>=0.0.5
python-dotenv>=1.0.0
pypdf>=3.17.0
python-docx>=1.1.0
numpy>=1.26.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pytest>=7.4.0
httpx>=0.25.0
